% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{caclust}
\alias{caclust}
\alias{caclust,cacomp-method}
\alias{caclust,SingleCellExperiment-method}
\title{caclust}
\usage{
caclust(
  obj,
  k,
  algorithm = "leiden",
  SNN_prune = 1/15,
  loops = FALSE,
  mode = "out",
  select_genes = TRUE,
  prune_overlap = TRUE,
  overlap = 0.2,
  calc_gene_cell_kNN = FALSE,
  resolution = 1,
  marker_genes = NULL,
  n.int = 10,
  rand_seed = 2358,
  use_gap = TRUE,
  nclust = NULL,
  spectral_method = "kmeans",
  iter_max = 10,
  num_seeds = 10,
  return_eig = TRUE,
  dims = NULL,
  cast_to_dense = TRUE,
  ...
)

\S4method{caclust}{cacomp}(
  obj,
  k,
  algorithm = "leiden",
  SNN_prune = 1/15,
  loops = FALSE,
  mode = "out",
  select_genes = TRUE,
  prune_overlap = TRUE,
  overlap = 0.2,
  calc_gene_cell_kNN = FALSE,
  resolution = 1,
  marker_genes = NULL,
  n.int = 10,
  rand_seed = 2358,
  use_gap = TRUE,
  nclust = NULL,
  spectral_method = "kmeans",
  iter_max = 10,
  num_seeds = 10,
  return_eig = TRUE,
  dims = NULL,
  cast_to_dense = TRUE,
  method = BiocNeighbors::KmknnParam(),
  BPPARAM = BiocParallel::SerialParam(),
  leiden_pack = "igraph",
  ...
)

\S4method{caclust}{SingleCellExperiment}(
  obj,
  k,
  algorithm = "leiden",
  SNN_prune = 1/15,
  loops = FALSE,
  mode = "out",
  select_genes = TRUE,
  prune_overlap = TRUE,
  overlap = 0.2,
  calc_gene_cell_kNN = FALSE,
  resolution = 1,
  marker_genes = NULL,
  n.int = 10,
  rand_seed = 2358,
  use_gap = TRUE,
  nclust = NULL,
  spectral_method = "kmeans",
  iter_max = 10,
  num_seeds = 10,
  return_eig = TRUE,
  dims = NULL,
  cast_to_dense = TRUE,
  method = BiocNeighbors::KmknnParam(),
  BPPARAM = BiocParallel::SerialParam(),
  leiden_pack = "igraph",
  ...,
  caclust_meta_name = "caclust",
  cacomp_meta_name = "CA"
)
}
\arguments{
\item{obj}{A cacomp object or SingleCellExperiment object}

\item{...}{further arguments}

\item{caclust_meta_name}{the name of caclust object stored in
metadata(SingleCellExperiment object). Default: 'caclust.'}

\item{cacomp_meta_name}{Character. The name of cacomp object stored in
metadata(SingleCellExperiment object). Default: 'caobj'.}
}
\value{
A caclust object or SingleCellExperiment object
}
\description{
`caclust()` performs biclustering on either a "cacomp" or
"SingleCellExperiment" object.
}
\details{
Convenient wrapper around `make_SNN` and `run_leiden`/`run_spectral`.
`run_caclust` takes a cacomp object and biclusters cells and genes.
}
\seealso{
Other biclustering: 
\code{\link{run_leiden}()},
\code{\link{run_spectral}()}
}
\concept{biclustering}
