% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{create_bipartite}
\alias{create_bipartite}
\title{Compute cell-gene bi-adjacency matrix}
\usage{
create_bipartite(
  caobj,
  k,
  min_edges = 0,
  MNN = FALSE,
  loops = FALSE,
  marker_genes = NULL,
  save_dists = TRUE,
  handle_isolated = c("remove", "connect_closest", "keep"),
  method = BiocNeighbors::KmknnParam(),
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{caobj}{A cacomp object with standard and principal coordinates
calculated.}

\item{k}{Integer. Number of genes each cell is connected to.
If MNN = TRUE it is additionally the number of edges between genes to cells.}

\item{MNN}{If TRUE a mutual nearest neighbours graph is computed.}

\item{marker_genes}{character. Optional. Names of known marker genes that
should be excempt from any pruning on the graph and be kept.}

\item{save_dists}{If TRUE saves distance matrix in the caclust object.}

\item{handle_isolated}{How to handle isolated cells
(cells with no gene connections). Can be either "remove",
"connect_closest" or "keep".}

\item{method}{\link[BiocNeighbors:BiocNeighborParam]{BiocNeighbors::BiocNeighborParam} object specifying the
algorithm to use. see Details.}

\item{BPPARAM}{\link{BiocParallel} settings parameter. By default single core
\code{\link[BiocParallel:SerialParam-class]{BiocParallel::SerialParam()}} but other parameters can be passed.}
}
\value{
Incidence matrix of type \code{dgCMatrix}.
}
\description{
Builds a single incidence matrix consisting of cells and genes in a bipartite graph.
}
\details{
\code{method} should be a kNN algorithm defined by
\link[BiocNeighbors:BiocNeighborParam]{BiocNeighbors::BiocNeighborParam}. For exact kNN search use
\code{BiocNeighbors::KmknnParam()} or \code{BiocNeighbors::VptreeParam()}.
}
