% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{run_caclust_bip}
\alias{run_caclust_bip}
\title{Run biclustering}
\usage{
run_caclust_bip(
  caobj,
  k,
  algorithm = "leiden",
  MNN = FALSE,
  resolution = 1,
  marker_genes = NULL,
  n.int = 10,
  rand_seed = 2358,
  use_gap = TRUE,
  nclust = NULL,
  spectral_method = "kmeans",
  iter_max = 10,
  num_seeds = 10,
  min_edges = 0,
  dims = NULL,
  save_dists = TRUE,
  handle_isolated = "remove",
  method = BiocNeighbors::KmknnParam(),
  BPPARAM = BiocParallel::SerialParam(),
  leiden_pack = "leiden",
  return_eig = FALSE
)
}
\arguments{
\item{caobj}{A cacomp object with principal and standard coordinates
calculated.}

\item{k}{Integer. Number of genes each cell is connected to.
If MNN = TRUE it is additionally the number of edges between genes to cells.}

\item{algorithm}{Character. Algorithm for clustering.
Options are "leiden" or "spectral". Defalut: 'leiden'.}

\item{MNN}{If TRUE a mutual nearest neighbours graph is computed.}

\item{resolution}{float number. Resolution for leiden algorithm.}

\item{marker_genes}{character. Optional. Names of known marker genes that
should be excempt from any pruning on the graph and be kept.}

\item{n.int}{Integer. Number of iterations for leiden algorithm.}

\item{rand_seed}{integer. Random seed.}

\item{use_gap}{Logical, TRUE/FALSE. If TRUE, 'eigengap' method will be used to find the
most important eigenvector automatically, and the number of output clusters
equals number of selected eigenvectors. If FALSE, 'nclust'(integer) should be specified.
The eigenvectors corresponding with the smallest 'nclust' eigenvalues will be
selcted and 'nclust' clusters will be detected by skmeans/kmeans/GMM.}

\item{nclust}{Integer. Number of clusters.}

\item{spectral_method}{character. Name of the method to cluster the eigenvectors.
Can be on of the following 3:
\itemize{
\item "kmeans": k-means clustering
\item "skmeans": spherical k-means clustering
\item "GMM": Gaussian-Mixture-Model fuzzy clustering.
}}

\item{iter_max}{Number of iterations for k-means clustering and GMM.}

\item{num_seeds}{Number of times k-means clustering is repeated.}

\item{min_edges}{Minimum number of edges per gene. If a gene has less edges
than min_edges it is filtered out.}

\item{dims}{Integer. Number of dimensions to choose from SVD of graph laplacian.}

\item{save_dists}{If TRUE saves distance matrix in the caclust object.}

\item{handle_isolated}{How to handle isolated cells
(cells with no gene connections). Can be either "remove",
"connect_closest" or "keep".}

\item{method}{\link[BiocNeighbors:BiocNeighborParam]{BiocNeighbors::BiocNeighborParam} object specifying the
algorithm to use. see Details.}

\item{BPPARAM}{\link{BiocParallel} settings parameter. By default single core
\code{\link[BiocParallel:SerialParam-class]{BiocParallel::SerialParam()}} but other parameters can be passed.}

\item{leiden_pack}{character. Optional values are 'igraph'(default) and 'leiden', the package used for leiden clustering.}

\item{return_eig}{Saves computed eigenvectors in the caclust object.}
}
\value{
Returns caclust objec with clustering results stored. The cell and gene
clusters can be accessed via \code{cell_clusters(obj)}/\code{gene_clusters(obj)}.
}
\description{
Convenient wrapper around \code{make_SNN} and \code{run_leiden}/\code{run_spectral}.
\code{run_caclust_bip} takes a cacomp object and biclusters cells and genes.
}
