% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biMAP.R
\name{biMAP}
\alias{biMAP}
\alias{biMAP,caclust-method}
\alias{biMAP,SingleCellExperiment-method}
\title{Compute biMAP}
\usage{
biMAP(
  obj,
  caobj,
  k_umap = 30,
  k_knn = min(c(Matrix::rowSums(obj@inc), Matrix::colSums(obj@inc)), na.rm = TRUE),
  rand_seed = 2358,
  method = "dist",
  use_inc = TRUE,
  features = NULL,
  ...
)

\S4method{biMAP}{caclust}(
  obj,
  caobj,
  k_umap = 30,
  k_knn = min(c(Matrix::rowSums(obj@inc), Matrix::colSums(obj@inc)), na.rm = TRUE),
  rand_seed = 2358,
  method = "dist",
  use_inc = TRUE,
  features = NULL,
  ...
)

\S4method{biMAP}{SingleCellExperiment}(
  obj,
  caobj,
  k_umap = 30,
  k_knn = min(c(Matrix::rowSums(obj@inc), Matrix::colSums(obj@inc)), na.rm = TRUE),
  rand_seed = 2358,
  method = "dist",
  use_inc = TRUE,
  features = NULL,
  ...,
  caclust_meta_name = "caclust"
)
}
\arguments{
\item{obj}{A caclust object or SingleCellExperiment object}

\item{caobj}{A cacomp object with principal and standard coordinates
calculated. Only needs to be supplied when using method "ca".}

\item{k_umap}{integer. Number of nearest neighbours to use to compute UMAP.}

\item{k_knn}{integer. Number of nearest neighbours to use use from the
bipartite graph for construciton of the pre-computed kNN graph for UMAP.}

\item{rand_seed}{integer. Random seed for UMAP.}

\item{method}{Can be either "dist" or "spectral".}

\item{use_inc}{TRUE/FALSE. This parameter only works when method == 'ca'.
If TRUE, it will calculate bimap embedding of genes and cells contained in
the bipartite graph together with the given 'features' if any.
If FALSE, only calculate bimap embedding of cells contained in the bipartite
graph together with the given 'fearures' if any.}

\item{features}{character/vector of gene names.
This parameter only works when method == 'ca'.
'ca' method allows visualization of genes which are absent from biparitte
graph, so the 'features' could be genes detected by our biclustering
algorithm, and genes which goes through CA analysis.
Setting use_inc = FALSE allows users to have a visualization of the
features genes user defined/interested.}

\item{...}{Further arguments}

\item{caclust_meta_name}{The name of caclust object stored in metadata
(SingleCellExperiment object).}
}
\value{
A caclust object or SingleCellExperiment object.
}
\description{
The function takes either a \code{caclust} or \code{SingleCellExperiment}as input and
stores the biMAP in the "bimap" slot in the caclust object. If a
SingleCellExperiment was provided the caclust object is stored in its
metadata.
}
\details{
The biMAP cell and gene embeddings can be calculated via different methods
as controlled by the parameter \code{method}:
\itemize{
\item 'dist'(Default): run UMAP on the distance matrix of cell-gene SNN graph
built up by caclust, which is '1-adj(SNN)'.
\item 'spectral': run UMAP on the selected eigenvectors of cell-gene graph
laplacian (only eligiable when algorithm is set as 'spectral' in 'caclust'
function)
}
}
